requirements.txt
praw==7.7.0
python-dotenv==1.0.0
huggingface_hub==0.16.4
tqdm==4.66.1
requests==2.31.0
beautifulsoup4==4.12.2